pyproject.toml
[project]
name = "resume-modifier-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amazon-textract-textractor>=1.9.2",
    "boto3>=1.42.80",
    "strands-agents>=1.34.0",
    "strands-agents-builder>=0.1.10",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.10",
    "moto>=5.1.22",
    "pytest>=9.0.2",
]
